phase3: split output functions into src/output.cpp
Extract Average, OutputField, OutputSnapshot, OutputVisual, OutputCHM,
OutputLAI, TrackingData_andOutput, all OutputABC*/UpdateABC* helpers,
MPI_Share* stubs, and CloseOutputs (~3187 lines) from troll.cpp.
