+++ CMakeLists.txt
@@ -26,6 +26,7 @@ add_executable(${PROJECT_NAME}
     src/lookup_tables.cpp
     src/input.cpp
     src/memory.cpp
+    src/output.cpp
     src/lookUpTables_cache.cpp
     src/params/param_registry.cpp
 )
